Work on MVP pattern --> Moved classes to appropriate packages. Bug fix --> PermissionManager line 87: call to ActivityFetcher.getContext was null. The activity 'ActivityFetcher' is never created in the application. Used getCurrentActivity() instead
