+++ app/src/main/java/ca/uqac/bigdataetmoi/IBasePresenter.java
@@ -1,0 +1,5 @@
+package ca.uqac.bigdataetmoi;
+
+public interface IBasePresenter {
+    void start();
+}
