+++ app/src/main/java/ca/uqac/bigdataetmoi/IBaseView.java
@@ -1,0 +1,5 @@
+package ca.uqac.bigdataetmoi;
+
+public interface IBaseView<T> {
+    public void setPresenter(T presenter);
+}
