+++ app/src/main/res/layout/activity_main.xml
@@ -16,7 +16,7 @@
             android:orientation="vertical"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context="ca.uqac.bigdataetmoi.MainActivity">
+            tools:context="ca.uqac.bigdataetmoi.startup.MainMenu">
 
             <TextView
                 android:layout_width="match_parent"
